requirements.txt
pandas==2.0.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
sqlalchemy==2.0.19
tabulate==0.9.0